Initial Union Rep job event
- Based off of the other 3, need to:
  1) Fill in the descriptions
  2) do the icon
  3) Add some kind of 'union card' item
    - Which will probably just be something with the Weingarten rights on 'em
  4) Finalize the loadout
    - Just based it off the Lawyer.  Thematically appropriate, probably can be better

Not much mechanically here as its primarily RP driven.  Cause minor headaches for the HoP, etc.

In the future I would like to have some way for it be randomly assigned a department to represent (what do you mean the borg's have a union?!), but having it only as a Cargo union rep. for now is already pretty funny.